requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.23.0

segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0
opencv-python-headless>=4.7.0
tqdm>=4.65.0

python-dotenv>=1.0.1